Split the class into files
